Update header.styl
The <label id="lang-select-label"> is overlapped by the <select>, which is set to opacity: 0 but still captures the mouse hover. That’s why the CSS :hover on #lang-select-label never triggers.
I fixed adding two rules to the botton of the css file and the same transition that the rest of the buttons on the nav have.

Signed-off-by: Matheus de Souza Santos  <86740915+Matt-develop@users.noreply.github.com>